Flattened delete command
